+++ src/frontend/app/src/routes/+page.svelte
@@ -1,0 +1,17 @@
+<script>
+    let title = "Variable title"
+</script>
+
+<div class="index">
+    <h1>{title}</h1>
+    <p>Some dummy text</p>
+</div>
+
+<style>
+    .index {
+        margin: 0px auto;
+        display: block;
+        text-align: center;
+        border: 1px red solid;
+    }
+</style>