+++ docs/user/changelog.md
@@ -1,4 +1,14 @@
 # CHANGELOG
+## version 0.0.2
+status: **IN PROGRESS**<br>
+date of release: **unknown**<br>
+content:
+### Added ###
+* As a user, I can create, read, and delete user groups by calling the IAM backend API
+* As a user, I can create, read, and delete users by calling the IAM backend API
+* As a user, I can update the rights of an existing group and those are propagated to all members
+* As a user, I can update the rights of an existing user
+* Authentication through JWT is activated (but we do not check permissions)
 ## version 0.0.1
 status: **OK**<br>
 date of release: **Aug 11, 2023**<br>